Move es util functions into es folder.
